class and instance methods with the same name
Have the New Relic Ruby agent report on code level metrics for two
separate methods with transaction traces. Both methods bear the same
name and are defined in the same class. One is an instance method and
one is a class (static) method.

When the demo runs, 2 new spans (1 for each method) will recorded within
the `agents#show` controller method transaction.

These spans will appear to be almost identical, with the `code.lineno`
value being the key CLM difference between the two.

The class is named `WhichIsWhich` and the methods are named `samename`.